fix grade
