+++ Omniscient-Back/src/main/java/com/omniscient/omniscientback/api/gradeapi/repository/GradeApiRepository.java
@@ -1,6 +1,6 @@
-package com.omniscient.omniscientback.api.gradejob.repository;
+package com.omniscient.omniscientback.api.gradeapi.repository;
 
-import com.omniscient.omniscientback.api.gradejob.model.GradeEntity;
+import com.omniscient.omniscientback.api.gradeapi.model.GradeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
